Make compiler happy
git-svn-id: https://proxytunnel.svn.sourceforge.net/svnroot/proxytunnel/trunk/proxytunnel@96 bc163920-b10d-0410-b2c5-a5491ca2ceef

+++ setproctitle.c
@@ -164,7 +164,7 @@ initsetproctitle(argc, argv, envp)
 		if (p == NULL)
 			return -1;
 		environ[i] = strncpy(p, envp[i], len);
-		p[len - 1] = NULL;
+		p[len - 1] = (int)NULL;
 	}
 	environ[i] = NULL;
